pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mircrew-indexer"
version = "1.0.0"
description = "Torznab-compatible indexer for MirCrew releases"
readme = "docs/README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
keywords = ["mircrew", "indexer", "torznab", "magnet", "torrent"]
authors = [
    {name = "MirCrew Community"}
]
maintainers = [
    {name = "MirCrew Community"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "python-dotenv>=0.19.0",
    "flask>=2.0.0",
    "lxml>=4.9.0",
]

[project.urls]
Homepage = "https://github.com/mircrew/mircrew-indexer"
Documentation = "https://github.com/mircrew/mircrew-indexer#readme"
Repository = "https://github.com/mircrew/mircrew-indexer.git"
Issues = "https://github.com/mircrew/mircrew-indexer/issues"
Changelog = "https://github.com/mircrew/mircrew-indexer/blob/main/CHANGELOG.md"

[project.scripts]
mircrew-indexer = "mircrew.core.indexer:main"
mircrew-api = "mircrew.api.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "bs4.*",
    "dotenv.*",
]
ignore_missing_imports = true